update code from calculator to App.js and left only the components
